Refactor main_gui.py to fix module-level execution and improve code health
- Moved `QT_LOGGING_RULES` environment variable setup and PyQt imports into `main()` function to prevent side effects on import.
- Extracted `TopLevelWindowLogger` and `WrappingSplashScreen` classes to new module `src/gui/startup.py`.
- Verified changes with import tests and existing functional tests.
